// src/app/shared/models/member-account.model.ts
export class MemberAccount {
  domain?: string;
  uuid?: string;
  account?: string;
  createDate?: Date;
  creatorId?: string;
  creatorName?: string;
  email?: string;
  lastModifyDate?: Date;
  modifierId?: string;
  modifierName?: string;
  name?: string;
  status?: string;
  type?: string;
  cardNo?: string;
  password?: string;
}